backend/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
pandas==2.2.3
reportlab==4.2.2
dj-database-url==2.2.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
